Value of url Neo4j connector
I make it more clear concerning the value of the Neo4j url connector. The original value "semantically" said nothing to help new users about what could be the value of the url to connect to Neo4j database. Questions were out there with no answers. It should be the path to the database. It's working with me and I hope it's for others. Feel free to correct me.